@@ -1,15 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.faa1192.weatherforecast"
         minSdkVersion 18
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 2
         versionName "0.2"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         resConfigs "en", "ru", "uk"
       //  resConfigs "nodpi", "hdpi", "xhdpi", "xxhdpi", "xxxhdpi"
         dataBinding.enabled = true
@@ -35,25 +34,26 @@ ext {
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile "com.android.support:appcompat-v7:27+"
-    compile "com.android.support:support-v4:27+"
-    compile "com.android.support:cardview-v7:27+"
-    compile "com.android.support:recyclerview-v7:27+"
-    compile "com.android.support:design:27+"
-    compile 'com.squareup.okhttp3:okhttp:3.4.1'
-    compile 'com.google.firebase:firebase-crash:10.0.1'
-    androidTestCompile "com.android.support:support-annotations:27+"
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation 'androidx.appcompat:appcompat:1.0.0-rc02'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0-rc02'
+    implementation 'androidx.cardview:cardview:1.0.0-rc02'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0-rc02'
+    implementation 'com.google.android.material:material:1.0.0-rc01'
+    implementation 'com.squareup.okhttp3:okhttp:3.11.0'
 
+    implementation 'com.google.firebase:firebase-crash:16.2.0'
+    androidTestImplementation 'androidx.annotation:annotation:1.0.0-rc02'
 
-    testCompile "junit:junit:4.12"
-    androidTestCompile "com.android.support.test.espresso:espresso-core:2.2.2"
-    androidTestCompile "com.android.support.test:runner:0.5"
-    compile 'com.android.support.constraint:constraint-layout:2.0.0-alpha1'
-    compile 'com.google.code.gson:gson:2.4'
-   // compile 'com.google.android.gms:play-services:10.0.1'
-    //compile 'com.google.android.gms:play-services-appindexing:10.0.1'
-    compile 'com.google.firebase:firebase-core:10.0.1'
-    compile 'com.google.firebase:firebase-appindexing:10.0.1'
+    implementation 'com.android.support:multidex:1.0.3'
+
+    testImplementation "junit:junit:4.12"
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0-alpha4'
+    androidTestImplementation 'androidx.test:runner:1.1.0-alpha4'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.google.code.gson:gson:2.8.2'
+  //  implementation 'com.google.android.gms:play-services:12.0.1'
+    implementation 'com.google.firebase:firebase-core:16.0.3'
+    implementation 'com.google.firebase:firebase-appindexing:16.0.1'
 }
 apply plugin: 'com.google.gms.google-services'
