@@ -3,8 +3,8 @@ package com.faa1192.weatherforecast.Countries;
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.Context;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
